requirements.txt
selenium
webdriver-manager
beautifulsoup4
requests
lxml
python-dotenv
elasticsearch
elasticsearch-dsl